Change handling of list of real-time systems
This marks a big behavior change in the module.

The antelope::startup defined type has had it's parameter rtsystems renamed to dirs. The dirs parameter can be either a string or an array. If it's a string, we try to split on a comma and assign that value to the variable $real_dirs. Due to this change, the template for the Antelope init script has had all instances of rtsystems changed to the variable real_dirs.

The main antelope module in init.pp now respects the rtsystems variable. This can be specified as a string, an array of strings, or a hash of hashes.