Add auto-depend on facts.d dir and ability to not manage fact
The concat instance now auto-requires the facts.d directory.

In order to make this module usable without the concat instance or
declaring the file resource, there is a new parameter for
antelope::instance called manage_fact that when false will not try to
create the antelope_instances fact.

For the top level antelope class, there is an parameter called
manage_instances_fact that controls the singleton antelope::instance's
fact creation