title can differ from version and explicit path
Add the ability to have a title for the resource separate from the
version, as well as the ability to specify an explicit path for the
file.
These changes will allow the license file for a particular version to
be put in both the proper $ANTELOPE/pf location plus multiple staging
areas, which will be useful for automated builds.