Add path param, allow title different from version
Allow us to specify a path independent from the version in use. Also
allow an arbitrary resource title that lets us use a version string.

These two changes will allow us to stage a site.pf to arbitrary
locations, such as a directory in a user's PFPATH or to a staging
location for automated builds.